refactored App and index.jsx to separate the render function and export function
